crossplatform dkim setup, Startup.cs cleanup, general cleanup
