--- ManualTester/ManualTester.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Bogus" Version="35.4.0" />
-    <PackageReference Include="Otp.NET" Version="1.3.0" />
-    <PackageReference Include="QRCoder" Version="1.4.3" />
-  </ItemGroup>
-
-</Project>
