+++ websmtp.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "websmtp", "websmtp\websmtp.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "tests\Tests.csproj", "{46E73802-BD95-4F8D-8B67-7568DA7E4635}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ManualTester", "ManualTester\ManualTester.csproj", "{4D545E43-22BA-474E-90B5-EE32A6835030}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -23,10 +21,6 @@ Global
 		{46E73802-BD95-4F8D-8B67-7568DA7E4635}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{46E73802-BD95-4F8D-8B67-7568DA7E4635}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{46E73802-BD95-4F8D-8B67-7568DA7E4635}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4D545E43-22BA-474E-90B5-EE32A6835030}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4D545E43-22BA-474E-90B5-EE32A6835030}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4D545E43-22BA-474E-90B5-EE32A6835030}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4D545E43-22BA-474E-90B5-EE32A6835030}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
